Use Patch operation for CnsFileAccessConfig updates
