+++ pkg/syncer/cnsoperator/util/util.go
@@ -18,7 +18,6 @@ package util
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -31,7 +30,6 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	apitypes "k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/strategicpatch"
 	"k8s.io/client-go/dynamic"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	cnsvsphere "sigs.k8s.io/vsphere-csi-driver/v3/pkg/common/cns-lib/vsphere"
@@ -398,47 +396,4 @@ func GetMaxWorkerThreads(ctx context.Context, key string, defaultVal int) int {
 			workerThreads)
 	}
 	return workerThreads
-}
-
-// PatchObject patches a Kubernetes object using strategic merge patch.
-// This function creates a patch between the original and modified objects and applies it using the client.
-// It returns an error if the patch operation fails.
-func PatchObject(ctx context.Context, k8sClient client.Client, original, modified client.Object) error {
-	log := logger.GetLogger(ctx)
-
-	// Marshal the original object
-	oldData, err := json.Marshal(original)
-	if err != nil {
-		log.Errorf("PatchObject: Failed to marshal original object %s/%s: %v",
-			original.GetNamespace(), original.GetName(), err)
-		return err
-	}
-
-	// Marshal the modified object
-	newData, err := json.Marshal(modified)
-	if err != nil {
-		log.Errorf("PatchObject: Failed to marshal modified object %s/%s: %v",
-			modified.GetNamespace(), modified.GetName(), err)
-		return err
-	}
-
-	// Create strategic merge patch
-	patchBytes, err := strategicpatch.CreateTwoWayMergePatch(oldData, newData, original)
-	if err != nil {
-		log.Errorf("PatchObject: Error creating strategic merge patch for object %s/%s: %v",
-			original.GetNamespace(), original.GetName(), err)
-		return err
-	}
-
-	// Apply the patch
-	patch := client.RawPatch(apitypes.StrategicMergePatchType, patchBytes)
-	if err := k8sClient.Patch(ctx, original, patch); err != nil {
-		log.Errorf("PatchObject: Failed to patch object %s/%s: %v",
-			original.GetNamespace(), original.GetName(), err)
-		return err
-	}
-
-	log.Debugf("PatchObject: Successfully patched object %s/%s",
-		original.GetNamespace(), original.GetName())
-	return nil
 }
